Move scale tests to scale directory
